requirements.txt
torch 
torchaudio 
numpy 
google-generativeai
sounddevice